+++ extensions/open-link-in-specific-browser/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Open Link in Specific Browser Changelog
 
+## [Hide Specific Browsers] - {PR_MERGE_DATE}
+
+- Add the ability to hide specific browsers from the list and the menu bar
+
 ## [New Command] - 2025-03-08
 
 - Add new command: Open selected text in default browser